inital commit
a lot of stuff was made before creating this repo just to test the waters.
and the waters seem good to surf.

(the code here was for testing. it can be expanded after figuring out how to associate the asset database to dlc blob files and how the compression works)